requirements-ubuntu.txt
numpy==1.16.4
pillow==6.1.0
opencv-python==4.1.0.25
rsa==4.0
https://download.pytorch.org/whl/cu100/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
torchvision==0.2.2.post3
imageio==2.5.0
python-dotenv==0.10.3
rook==0.1.73
